test: 이용 약관 모달 스토리 수정
